update phần admin và desk
