+++ src/main/java/com/codegym/repository/OrderDetailRepository.java
@@ -4,6 +4,10 @@ import com.codegym.entity.OrderDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
+
 @Repository
 public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+    List<OrderDetail> findOrderDetailByOrder_Id(Long id);
 }
