+++ src/main/java/com/codegym/service/orderDetail/IOrderDetailService.java
@@ -3,5 +3,9 @@ package com.codegym.service.orderDetail;
 import com.codegym.entity.OrderDetail;
 import com.codegym.service.IGeneralService;
 
+import java.util.List;
+
 public interface IOrderDetailService extends IGeneralService<OrderDetail> {
+
+    List <OrderDetail> findOrderDetailByOrder_id(Long id);
 }
