+++ src/main/java/com/codegym/service/orderDetail/OrderDetailService.java
@@ -5,6 +5,7 @@ import com.codegym.repository.OrderDetailRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -30,5 +31,10 @@ public class OrderDetailService implements IOrderDetailService {
     @Override
     public void remove(Long id) {
         orderDetailRepository.deleteById(id);
+    }
+
+    @Override
+    public List<OrderDetail> findOrderDetailByOrder_id(Long id) {
+        return orderDetailRepository.findOrderDetailByOrder_Id(id);
     }
 }
