+++ src/main/webapp/WEB-INF/views/layout/desk.html
@@ -16,7 +16,6 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.2/dist/js/bootstrap.bundle.min.js"></script>
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css"
           integrity="sha384-9gVQ4dYFwwWSjIDZnLEWnxCjeSWFphJiwGPXr1jddIhOegiu1FwO5qRGvFXOdJZ4" crossorigin="anonymous">
-
 </head>
 <body>
 <!--preloader-->
@@ -76,16 +75,17 @@
                 <form id="deskCreate">
                     <div class="form-group">
                         <label for="deskName">Tên Bàn</label>
-                        <input type="text" class="form-control" id="deskName" data-rule-required="true" data-msg-required="Không được bỏ trống">
+                        <input type="text" class="form-control" id="deskName" data-rule-required="true"
+                               data-msg-required="Không được bỏ trống">
                     </div>
                     <button type="button" class="btn btn-danger" onclick="createTable()">Tạo Bàn</button>
                 </form>
             </div>
             <div class="col-xl-6">
-                    <div class="desk-list">
-                        <div class="row">
-                        </div>
+                <div class="desk-list">
+                    <div class="row">
                     </div>
+                </div>
             </div>
             <!--   Big container   -->
 
@@ -95,15 +95,46 @@
                     <div class="card wizard-card" data-color="red" id="wizard">
                         <div class="wizard-header">
                             <h2 class="wizard-title" style="color: #00A0D1">
-                                Thông tin bàn
+                                Thông tin bàn đang phục vụ
                             </h2>
-                        </div>
-                        <div class="wizard-navigation">
-                            <ul>
-                            </ul>
-                        </div>
-                        <div class="tab-content">
+                            <div class="wizard-navigation">
+                                <ul>
 
+                                </ul>
+                            </div>
+
+                            <div class="tab-content">
+                                <div class="tab-pane" id="table1">
+                                    <div class="row">
+                                        <h1>bàn 1</h1>
+                                    </div>
+                                </div>
+                                <div class="tab-pane" id="table2">
+                                    <div class="row">
+                                        <h1>bàn 2</h1>
+                                    </div>
+                                </div>
+                                <div class="tab-pane" id="table3">
+                                    <div class="row">
+                                        <h1>bàn 3</h1>
+                                    </div>
+                                </div>
+                                <div class="tab-pane" id="table4">
+                                    <div class="row">
+                                        <h1>bàn 4</h1>
+                                    </div>
+                                </div>
+                                <div class="tab-pane" id="table5">
+                                    <div class="row">
+                                        <h1>bàn 5</h1>
+                                    </div>
+                                </div>
+                                <div class="tab-pane" id="table6">
+                                    <div class="row">
+                                        <h1>bàn 6</h1>
+                                    </div>
+                                </div>
+                            </div>
                         </div>
                     </div>
                 </div> <!-- wizard container -->
@@ -133,8 +164,8 @@
 
 <!--   Core JS Files-->
 
-<script src="/resource/js/desk/desk.js"></script>
-<!--   Core JS Files-->
+<script src="/resource/js/deskPage/desk.js"></script>
+<!--  Core JS Files-->
 <script src="/resource/assets-desk/js/jquery-2.2.4.min.js" type="text/javascript"></script>
 <script src="/resource/assets-desk/js/bootstrap.min.js" type="text/javascript"></script>
 <script src="/resource/assets-desk/js/jquery.bootstrap.js" type="text/javascript"></script>
@@ -160,5 +191,6 @@
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js"
         integrity="sha384-uefMccjFJAIv6A+rW+L4AHf99KvxDjWSu1z9VI8SKNVmz4sk7buKt/6v9KI65qnm"
         crossorigin="anonymous"></script>
+
 </body>
 </html>
