+++ src/main/webapp/WEB-INF/views/user/home.html
@@ -82,35 +82,37 @@
             <div class="col-3">
                 <div class="card">
                     <div class="card-body">
-                           <form id="createOrderDetail" class="form-group">
-                               <label for="order_id"></label>
-                               <input type="number" id="order_id" hidden>
-                               <div class="form-group">
-                                   <label for="desk-order"><h2>Chọn Bàn</h2></label>
-                                   <select class="form-control" id="desk-order">
-                                        <!--Danh sách để chọn bàn order-->
-                                   </select>
-                               </div>
-                               <div class="form-group">
-                                   <label for="product-order"><h3>Chọn món</h3></label>
-                                   <select class="form-control" id="product-order">
-                                       <!--Danh sách để chọn món-->
-                                   </select>
-                               </div>
-                               <div class="form-group">
-                                   <label for="product-price"><h3>Giá</h3></label>
-                                   <span><input class="form-control" id="product-price" type="number" readonly> vnđ </span>
-                               </div>
-                               <div class="form-group">
-                                   <label for="product-quantity"><h3>Số lượng</h3></label>
-                                   <input class="form-control" id="product-quantity" type="number" data-rule-min="0" data-msg-required="Không được bỏ trống">
-                               </div>
-                               <div class="form-group">
-                                   <label for="product-unitPrice"><h3>Thành tiền: </h3></label>
-                                   <input class="form-control" id="product-unitPrice" type="number" readonly>
-                               </div>
-                               <button onclick="createOrderDetail()" type="button" class="btn btn-success">Nhập món</button>
-                           </form>
+                        <form id="createOrderDetail" class="form-group">
+                            <label for="order_id"></label>
+                            <input type="number" id="order_id" hidden>
+                            <div class="form-group">
+                                <label for="desk-order"><h2>Chọn Bàn</h2></label>
+                                <select class="form-control" id="desk-order">
+                                    <!--Danh sách để chọn bàn order-->
+                                </select>
+                            </div>
+                            <div class="form-group">
+                                <label for="product-order"><h3>Chọn món</h3></label>
+                                <select class="form-control" id="product-order">
+                                    <!--Danh sách để chọn món-->
+                                </select>
+                            </div>
+                            <div class="form-group">
+                                <label for="product-price"><h3>Giá</h3></label>
+                                <span><input class="form-control" id="product-price" type="number" readonly> vnđ </span>
+                            </div>
+                            <div class="form-group">
+                                <label for="product-quantity"><h3>Số lượng</h3></label>
+                                <input class="form-control" id="product-quantity" type="number" data-rule-min="0"
+                                       data-msg-required="Không được bỏ trống">
+                            </div>
+                            <div class="form-group">
+                                <label for="product-unitPrice"><h3>Thành tiền: </h3></label>
+                                <input class="form-control" id="product-unitPrice" type="number" readonly>
+                            </div>
+                            <button onclick="createOrderDetail()" type="button" class="btn btn-success">Nhập món
+                            </button>
+                        </form>
                     </div>
                 </div>
             </div>
@@ -153,14 +155,12 @@
 
                             </tbody>
                         </table>
-
                     </div> <!-- end card body-->
                 </div> <!-- end card -->
             </div><!-- end col-->
         </div>
     </div>
 </div>
-
 <!-- Hidden Navigation Bar -->
 <section class="hidden-bar right-align">
     <div class="hidden-bar-closer">
